requirements.txt
# DEAI pipeline — pip requirements (mirrors environment.yml for pip-only installs)
numpy>=1.26,<2.0
pandas>=2.2,<3.0
scipy>=1.13,<2.0
scikit-learn>=1.5,<2.0
statsmodels>=0.14,<1.0
matplotlib>=3.9,<4.0
seaborn>=0.13,<1.0
plotly>=5.22,<6.0
jupyterlab>=4.0,<5.0
pyyaml>=6.0
openpyxl>=3.1
xlrd>=2.0
requests>=2.31
tqdm>=4.66
joblib>=1.4
pyarrow>=16.0
fastparquet>=2024.2
pytest>=8.0
pytest-cov>=5.0
black>=24.0
isort>=5.13
xgboost>=2.0,<3.0
lightgbm>=4.3,<5.0
shap>=0.45,<1.0
lifelines>=0.28,<1.0
GEOparse>=2.0,<3.0
pydeseq2>=0.4,<1.0
gseapy>=1.1,<2.0
missingno>=0.5
tableone>=0.8
dcurves>=0.5
prince>=0.12
factor-analyzer>=0.5
adjustText>=1.0
kaleido>=0.2
